requirements.txt
pandas==1.3.3
streamlit==0.89.0
google-cloud-storage==1.41.0
pyomo==5.7
plotly==4.12.0
beautifulsoup4==4.9.3
matplotlib==3.3.2
openpyxl==3.0.4
xlrd==1.2.0
numpy-financial==1.0.0
tables==3.6.1
XlsxWriter==1.3.7
